requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.15.0=pypi_0
apache-beam=2.36.0=pypi_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h2bbff1b_0
astunparse=1.6.3=pypi_0
attrs=21.4.0=pyhd3eb1b0_0
avro-python3=1.10.2=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
bleach=4.1.0=pyhd3eb1b0_0
ca-certificates=2022.2.1=haa95532_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=pypi_0
cffi=1.15.0=py39h2bbff1b_1
charset-normalizer=2.0.12=pypi_0
cloudpickle=2.0.0=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
contextlib2=21.6.0=pypi_0
crcmod=1.7=pypi_0
cycler=0.11.0=pypi_0
cython=0.29.27=pypi_0
debugpy=1.5.1=py39hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dill=0.3.1.1=pypi_0
dm-tree=0.1.6=pypi_0
docopt=0.6.2=pypi_0
entrypoints=0.3=py39haa95532_0
fastavro=1.4.9=pypi_0
flatbuffers=1.12=pypi_0
fonttools=4.29.1=pypi_0
gast=0.4.0=pypi_0
gin-config=0.5.0=pypi_0
git=2.34.1=haa95532_0
google-api-core=2.5.0=pypi_0
google-api-python-client=2.37.0=pypi_0
google-auth=2.6.0=pypi_0
google-auth-httplib2=0.1.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
googleapis-common-protos=1.54.0=pypi_0
grpcio=1.34.1=pypi_0
h5py=3.1.0=pypi_0
hdfs=2.6.0=pypi_0
httplib2=0.19.1=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.11.1=pypi_0
importlib_metadata=4.8.2=hd3eb1b0_0
ipykernel=6.4.1=py39haa95532_1
ipython=7.31.1=py39haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.1=py39haa95532_1
jinja2=3.0.2=pyhd3eb1b0_0
joblib=1.1.0=pypi_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_core=4.9.1=py39haa95532_0
jupyterlab_pygments=0.1.2=py_0
kaggle=1.5.12=pypi_0
keras=2.8.0=pypi_0
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.2=pypi_0
libclang=13.0.0=pypi_0
lvis=0.5.3=pypi_0
lxml=4.8.0=pypi_0
markdown=3.3.6=pypi_0
markupsafe=2.0.1=py39h2bbff1b_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h2bbff1b_1000
nb_conda_kernels=2.3.1=py39haa95532_0
nbclient=0.5.11=pyhd3eb1b0_0
nbconvert=6.1.0=py39haa95532_0
nbformat=5.1.3=pyhd3eb1b0_0
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.4.8=py39haa95532_0
numpy=1.21.5=pypi_0
oauth2client=4.1.3=pypi_0
oauthlib=3.2.0=pypi_0
object-detection=0.1=pypi_0
opencv-python=4.5.5.62=pypi_0
opencv-python-headless=4.5.5.62=pypi_0
openssl=1.1.1m=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
orjson=3.6.7=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=pypi_0
pip=21.2.4=py39haa95532_0
portalocker=2.4.0=pypi_0
prometheus_client=0.13.1=pyhd3eb1b0_0
promise=2.3=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
proto-plus=1.20.3=pypi_0
protobuf=3.19.4=pypi_0
psutil=5.9.0=pypi_0
py-cpuinfo=8.0.0=pypi_0
pyarrow=6.0.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydot=1.4.2=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pymongo=3.12.3=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.18.0=py39h196d8e1_0
python=3.9.7=h6244533_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-slugify=6.0.1=pypi_0
pytz=2021.3=pypi_0
pywin32=303=pypi_0
pywinpty=2.0.2=py39h5da7b33_0
pyyaml=5.4.1=pypi_0
pyzmq=22.3.0=py39hd77b12b_2
regex=2022.1.18=pypi_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
sacrebleu=2.0.0=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
sentencepiece=0.1.96=pypi_0
seqeval=1.2.2=pypi_0
setuptools=60.9.1=pypi_0
six=1.15.0=pypi_0
sqlite=3.37.2=h2bbff1b_0
tabulate=0.8.9=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.8.0=pypi_0
tensorflow-addons=0.16.1=pypi_0
tensorflow-datasets=4.5.2=pypi_0
tensorflow-estimator=2.5.0=pypi_0
tensorflow-hub=0.12.0=pypi_0
tensorflow-io=0.24.0=pypi_0
tensorflow-io-gcs-filesystem=0.24.0=pypi_0
tensorflow-metadata=1.6.0=pypi_0
tensorflow-model-optimization=0.7.1=pypi_0
tensorflow-text=2.8.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.1=py39haa95532_0
testpath=0.5.0=pyhd3eb1b0_0
text-unidecode=1.3=pypi_0
tf-estimator-nightly=2.8.0.dev2021122109=pypi_0
tf-models-official=2.8.0=pypi_0
tf-slim=1.1.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tornado=6.1=py39h2bbff1b_0
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
typeguard=2.13.3=pypi_0
typing-extensions=3.7.4.3=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
uritemplate=4.1.1=pypi_0
urllib3=1.26.8=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39haa95532_1
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pypi_0
wincertstore=0.2=py39haa95532_2
winpty=0.4.3=4
wrapt=1.12.1=pypi_0
zipp=3.7.0=pyhd3eb1b0_0
